@@ -118,7 +118,7 @@ html_static_path = ['_static']
 # Output file base name for HTML help builder.
 htmlhelp_basename = 'pygame_gui_doc'
 
-
+smartquotes = False
 # -- Options for LaTeX output ------------------------------------------------
 
 latex_elements = {
