@@ -5,16 +5,16 @@ from pygame_gui.elements.ui_vertical_scroll_bar import UIVerticalScrollBar
 from pygame_gui.elements.ui_horizontal_scroll_bar import UIHorizontalScrollBar
 from pygame_gui.elements.ui_label import UILabel
 from pygame_gui.elements.ui_panel import UIPanel
+from pygame_gui.elements.ui_progress_bar import UIProgressBar
 from pygame_gui.elements.ui_screen_space_health_bar import UIScreenSpaceHealthBar
 from pygame_gui.elements.ui_selection_list import UISelectionList
 from pygame_gui.elements.ui_text_box import UITextBox
+from pygame_gui.elements.ui_text_entry_line import UITextEntryLine
 from pygame_gui.elements.ui_tool_tip import UITooltip
 from pygame_gui.elements.ui_drop_down_menu import UIDropDownMenu
 from pygame_gui.elements.ui_world_space_health_bar import UIWorldSpaceHealthBar
-from pygame_gui.elements.ui_text_entry_line import UITextEntryLine
 from pygame_gui.elements.ui_window import UIWindow
 from pygame_gui.elements.ui_scrolling_container import UIScrollingContainer
-from pygame_gui.elements.text import TextHTMLParser, FadeInEffect, FadeOutEffect, TypingAppearEffect
 
 
 __all__ = ['UIImage',
@@ -30,10 +30,7 @@ __all__ = ['UIImage',
            'UITooltip',
            'UIDropDownMenu',
            'UIWorldSpaceHealthBar',
+           'UIProgressBar',
            'UITextEntryLine',
            'UIWindow',
-           'UIScrollingContainer',
-           'TextHTMLParser',
-           'FadeInEffect',
-           'FadeOutEffect',
-           'TypingAppearEffect']
+           'UIScrollingContainer']
