@@ -3,11 +3,12 @@ from typing import Union, List, Tuple, Dict
 import pygame
 
 from pygame_gui._constants import UI_BUTTON_PRESSED, UI_SELECTION_LIST_NEW_SELECTION
-from pygame_gui._constants import UI_DROP_DOWN_MENU_CHANGED
+from pygame_gui._constants import UI_DROP_DOWN_MENU_CHANGED, OldType
 
 from pygame_gui.core.interfaces import IContainerLikeInterface, IUIManagerInterface
 from pygame_gui.core import UIElement, ObjectID
 from pygame_gui.core.drawable_shapes import RectDrawableShape, RoundedRectangleShape
+from pygame_gui.core.ui_container import UIContainer
 
 from pygame_gui.elements.ui_button import UIButton
 from pygame_gui.elements.ui_selection_list import UISelectionList
@@ -69,6 +70,8 @@ class UIExpandedDropDownState:
         self.should_transition = False
         self.target_state = 'closed'
 
+        self.active_buttons = []
+
     def rebuild(self):
         """
         Rebuild the state from theming parameters and dimensions.
@@ -105,7 +108,7 @@ class UIExpandedDropDownState:
                     expand_button_symbol = '▼'
             self.close_button.set_text(expand_button_symbol)
 
-    def start(self):
+    def start(self, should_rebuild: bool = True):
         """
         Called each time we enter the expanded state. It creates the necessary elements, the
         selected option, all the other available options and the close button.
@@ -113,7 +116,10 @@ class UIExpandedDropDownState:
         """
         self.should_transition = False
 
-        self.selected_option_button = UIButton(pygame.Rect(self.base_position_rect.topleft,
+        border_and_shadow = (self.drop_down_menu_ui.shadow_width +
+                             self.drop_down_menu_ui.border_width)
+        self.active_buttons = []
+        self.selected_option_button = UIButton(pygame.Rect((border_and_shadow, border_and_shadow),
                                                            (self.base_position_rect.width -
                                                             self.close_button_width,
                                                             self.base_position_rect.height)),
@@ -124,6 +130,7 @@ class UIExpandedDropDownState:
                                                parent_element=self.drop_down_menu_ui,
                                                object_id=ObjectID('#selected_option', None))
         self.drop_down_menu_ui.join_focus_sets(self.selected_option_button)
+        self.active_buttons.append(self.selected_option_button)
 
         expand_button_symbol = '▼'
 
@@ -139,28 +146,7 @@ class UIExpandedDropDownState:
                                                                        list_class_ids,
                                                                        list_object_ids)
 
-        try:
-            list_shadow_width = int(
-                self.ui_manager.get_theme().get_misc_data('shadow_width', final_ids))
-        except (LookupError, ValueError):
-            list_shadow_width = 2
-
-        try:
-            list_border_width = int(
-                self.ui_manager.get_theme().get_misc_data('border_width', final_ids))
-        except (LookupError, ValueError):
-            list_border_width = 1
-
-        try:
-            list_item_height = int(
-                self.ui_manager.get_theme().get_misc_data('list_item_height', final_ids))
-        except (LookupError, ValueError):
-            list_item_height = 20
-
-        options_list_border_and_shadow = list_shadow_width + list_border_width
-        self.options_list_height = ((list_item_height * len(self.options_list)) +
-                                    (2 * options_list_border_and_shadow))
-        self.option_list_y_pos = 0
+        self._calculate_options_list_sizes(final_ids)
         if self.expand_direction is not None:
             if self.expand_direction == 'up':
                 expand_button_symbol = '▲'
@@ -186,38 +172,62 @@ class UIExpandedDropDownState:
 
                 self.option_list_y_pos = self.base_position_rect.bottom
 
-        close_button_x = (self.base_position_rect.x +
-                          self.base_position_rect.width -
-                          self.close_button_width)
+        if self.close_button_width > 0:
+            close_button_x = (border_and_shadow +
+                              self.base_position_rect.width -
+                              self.close_button_width)
 
-        self.close_button = UIButton(pygame.Rect((close_button_x,
-                                                  self.base_position_rect.y),
-                                                 (self.close_button_width,
-                                                  self.base_position_rect.height)),
-                                     expand_button_symbol,
-                                     self.ui_manager,
-                                     self.ui_container,
-                                     starting_height=2,
-                                     parent_element=self.drop_down_menu_ui,
-                                     object_id='#expand_button')
-        self.drop_down_menu_ui.join_focus_sets(self.close_button)
-
+            self.close_button = UIButton(pygame.Rect((close_button_x,
+                                                      border_and_shadow),
+                                                     (self.close_button_width,
+                                                      self.base_position_rect.height)),
+                                         expand_button_symbol,
+                                         self.ui_manager,
+                                         self.ui_container,
+                                         starting_height=2,
+                                         parent_element=self.drop_down_menu_ui,
+                                         object_id='#expand_button')
+            self.drop_down_menu_ui.join_focus_sets(self.close_button)
+            self.active_buttons.append(self.close_button)
         list_rect = pygame.Rect(self.drop_down_menu_ui.relative_rect.left,
                                 self.option_list_y_pos,
                                 (self.drop_down_menu_ui.relative_rect.width -
                                  self.close_button_width),
                                 self.options_list_height)
         self.options_selection_list = UISelectionList(list_rect,
-                                                      starting_height=2,
+                                                      starting_height=3,
                                                       item_list=self.options_list,
                                                       allow_double_clicks=False,
                                                       manager=self.ui_manager,
                                                       parent_element=self.drop_down_menu_ui,
-                                                      container=self.ui_container,
+                                                      container=self.drop_down_menu_ui.ui_container,
+                                                      anchors=self.drop_down_menu_ui.anchors,
                                                       object_id='#drop_down_options_list')
         self.drop_down_menu_ui.join_focus_sets(self.options_selection_list)
 
-        self.rebuild()
+        if should_rebuild:
+            self.rebuild()
+
+    def _calculate_options_list_sizes(self, final_ids):
+        try:
+            list_shadow_width = int(
+                self.ui_manager.get_theme().get_misc_data('shadow_width', final_ids))
+        except (LookupError, ValueError):
+            list_shadow_width = 2
+        try:
+            list_border_width = int(
+                self.ui_manager.get_theme().get_misc_data('border_width', final_ids))
+        except (LookupError, ValueError):
+            list_border_width = 1
+        try:
+            list_item_height = int(
+                self.ui_manager.get_theme().get_misc_data('list_item_height', final_ids))
+        except (LookupError, ValueError):
+            list_item_height = 20
+        options_list_border_and_shadow = list_shadow_width + list_border_width
+        self.options_list_height = ((list_item_height * len(self.options_list)) +
+                                    (2 * options_list_border_and_shadow))
+        self.option_list_y_pos = 0
 
     def finish(self):
         """
@@ -225,7 +235,8 @@ class UIExpandedDropDownState:
         """
         self.options_selection_list.kill()
         self.selected_option_button.kill()
-        self.close_button.kill()
+        if self.close_button is not None:
+            self.close_button.kill()
 
     def process_event(self, event: pygame.event.Event) -> bool:
         """
@@ -237,25 +248,27 @@ class UIExpandedDropDownState:
                  rest of the UI.
 
         """
-        if (event.type == pygame.USEREVENT and
-                event.user_type == UI_BUTTON_PRESSED and
-                event.ui_element in [self.close_button, self.selected_option_button]):
-
+        if event.type == UI_BUTTON_PRESSED and event.ui_element in self.active_buttons:
             self.should_transition = True
 
-        if (event.type == pygame.USEREVENT and
-                event.user_type == UI_SELECTION_LIST_NEW_SELECTION and
+        if (event.type == UI_SELECTION_LIST_NEW_SELECTION and
                 event.ui_element == self.options_selection_list):
             selection = self.options_selection_list.get_single_selection()
             self.drop_down_menu_ui.selected_option = selection
             self.should_transition = True
 
-            event_data = {'user_type': UI_DROP_DOWN_MENU_CHANGED,
+            # old event - to be removed in 0.8.0
+            event_data = {'user_type': OldType(UI_DROP_DOWN_MENU_CHANGED),
                           'text': self.drop_down_menu_ui.selected_option,
                           'ui_element': self.drop_down_menu_ui,
                           'ui_object_id': self.drop_down_menu_ui.most_specific_combined_id}
-            drop_down_changed_event = pygame.event.Event(pygame.USEREVENT, event_data)
-            pygame.event.post(drop_down_changed_event)
+            pygame.event.post(pygame.event.Event(pygame.USEREVENT, event_data))
+
+            # new event
+            event_data = {'text': self.drop_down_menu_ui.selected_option,
+                          'ui_element': self.drop_down_menu_ui,
+                          'ui_object_id': self.drop_down_menu_ui.most_specific_combined_id}
+            pygame.event.post(pygame.event.Event(UI_DROP_DOWN_MENU_CHANGED, event_data))
 
         return False  # don't consume any events
 
@@ -274,14 +287,15 @@ class UIExpandedDropDownState:
         self.base_position_rect.y = self.drop_down_menu_ui.relative_rect.y + border_and_shadow
 
         # update all the ui elements that depend on the base position
-        self.selected_option_button.set_relative_position(self.base_position_rect.topleft)
+        self.selected_option_button.set_relative_position((border_and_shadow, border_and_shadow))
         list_post = (self.drop_down_menu_ui.relative_rect.left, self.option_list_y_pos)
         self.options_selection_list.set_relative_position(list_post)
 
-        close_button_x = (self.base_position_rect.x +
-                          self.base_position_rect.width -
-                          self.close_button_width)
-        self.close_button.set_relative_position([close_button_x, self.base_position_rect.y])
+        if self.close_button is not None:
+            close_button_x = (border_and_shadow +
+                              self.base_position_rect.width -
+                              self.close_button_width)
+            self.close_button.set_relative_position([close_button_x, border_and_shadow])
 
     def update_dimensions(self):
         """
@@ -320,15 +334,13 @@ class UIExpandedDropDownState:
 
         list_pos = (self.drop_down_menu_ui.relative_rect.left, self.option_list_y_pos)
         self.options_selection_list.set_relative_position(list_pos)
-
-        close_button_x = (self.base_position_rect.x +
-                          self.base_position_rect.width -
-                          self.close_button_width)
-        self.close_button.set_dimensions((self.close_button_width,
-                                          self.base_position_rect.height))
-        self.close_button.set_relative_position((close_button_x, self.base_position_rect.y))
-
-        # self.rebuild()
+        if self.close_button is not None:
+            close_button_x = (border_and_shadow +
+                              self.base_position_rect.width -
+                              self.close_button_width)
+            self.close_button.set_dimensions((self.close_button_width,
+                                              self.base_position_rect.height))
+            self.close_button.set_relative_position((close_button_x, border_and_shadow))
 
     def on_fresh_drawable_shape_ready(self):
         """
@@ -337,8 +349,7 @@ class UIExpandedDropDownState:
 
         In this case the result is to set the UI element's image to the new surface.
         """
-        image = self.drop_down_menu_ui.drawable_shape.get_fresh_surface()
-        self.drop_down_menu_ui.set_image(image)
+        self.drop_down_menu_ui.set_image(self.drop_down_menu_ui.drawable_shape.get_fresh_surface())
 
     def hide(self):
         """
@@ -394,12 +405,15 @@ class UIClosedDropDownState:
         self.target_state = 'expanded'
         self.visible = visible
 
+        self.active_buttons = []
+
     def disable(self):
         """
         Disables the closed state so that it is no longer interactive.
         """
         self.selected_option_button.disable()
-        self.open_button.disable()
+        if self.open_button is not None:
+            self.open_button.disable()
         self.drop_down_menu_ui.drawable_shape.set_active_state('disabled')
 
     def enable(self):
@@ -407,7 +421,8 @@ class UIClosedDropDownState:
         Re-enables the closed state so we can once again interact with it.
         """
         self.selected_option_button.enable()
-        self.open_button.enable()
+        if self.open_button is not None:
+            self.open_button.enable()
         self.drop_down_menu_ui.drawable_shape.set_active_state('normal')
 
     def rebuild(self):
@@ -435,7 +450,7 @@ class UIClosedDropDownState:
                                                                           ['normal', 'disabled'],
                                                                           self.ui_manager)
 
-        self.drop_down_menu_ui.image = self.drop_down_menu_ui.drawable_shape.get_fresh_surface()
+        self.drop_down_menu_ui.set_image(self.drop_down_menu_ui.drawable_shape.get_fresh_surface())
 
         # extra
         if self.open_button is not None:
@@ -447,16 +462,20 @@ class UIClosedDropDownState:
                     expand_button_symbol = '▼'
             self.open_button.set_text(expand_button_symbol)
 
-    def start(self):
+    def start(self, should_rebuild: bool = True):
         """
         Called each time we enter the closed state. It creates the necessary elements, the
         selected option and the open button.
         """
-        self.rebuild()
+        if should_rebuild:
+            self.rebuild()
 
         self.should_transition = False
-        self.selected_option_button = UIButton(pygame.Rect((self.base_position_rect.x,
-                                                            self.base_position_rect.y),
+
+        border_and_shadow = (self.drop_down_menu_ui.shadow_width +
+                             self.drop_down_menu_ui.border_width)
+        self.active_buttons = []
+        self.selected_option_button = UIButton(pygame.Rect((border_and_shadow, border_and_shadow),
                                                            (self.base_position_rect.width -
                                                             self.open_button_width,
                                                             self.base_position_rect.height)),
@@ -468,37 +487,39 @@ class UIClosedDropDownState:
                                                object_id='#selected_option',
                                                visible=self.visible)
         self.drop_down_menu_ui.join_focus_sets(self.selected_option_button)
+        self.active_buttons.append(self.selected_option_button)
 
-        open_button_x = (self.base_position_rect.x +
-                         self.base_position_rect.width -
-                         self.open_button_width)
-
-        expand_button_symbol = '▼'
-        if self.expand_direction is not None:
-            if self.expand_direction == 'up':
-                expand_button_symbol = '▲'
-            elif self.expand_direction == 'down':
-                expand_button_symbol = '▼'
-
-        self.open_button = UIButton(pygame.Rect((open_button_x,
-                                                 self.base_position_rect.y),
-                                                (self.open_button_width,
-                                                 self.base_position_rect.height)),
-                                    expand_button_symbol,
-                                    self.ui_manager,
-                                    self.ui_container,
-                                    starting_height=2,
-                                    parent_element=self.drop_down_menu_ui,
-                                    object_id='#expand_button',
-                                    visible=self.visible)
-        self.drop_down_menu_ui.join_focus_sets(self.open_button)
+        if self.open_button_width > 0:
+            open_button_x = (border_and_shadow +
+                             self.base_position_rect.width -
+                             self.open_button_width)
+            expand_button_symbol = '▼'
+            if self.expand_direction is not None:
+                if self.expand_direction == 'up':
+                    expand_button_symbol = '▲'
+                elif self.expand_direction == 'down':
+                    expand_button_symbol = '▼'
+            self.open_button = UIButton(pygame.Rect((open_button_x,
+                                                     border_and_shadow),
+                                                    (self.open_button_width,
+                                                     self.base_position_rect.height)),
+                                        expand_button_symbol,
+                                        self.ui_manager,
+                                        self.ui_container,
+                                        starting_height=2,
+                                        parent_element=self.drop_down_menu_ui,
+                                        object_id='#expand_button',
+                                        visible=self.visible)
+            self.drop_down_menu_ui.join_focus_sets(self.open_button)
+            self.active_buttons.append(self.open_button)
 
     def finish(self):
         """
         Called when we leave the closed state. Kills the open button and the selected option button.
         """
         self.selected_option_button.kill()
-        self.open_button.kill()
+        if self.open_button is not None:
+            self.open_button.kill()
 
     def process_event(self, event: pygame.event.Event) -> bool:
         """
@@ -509,9 +530,7 @@ class UIClosedDropDownState:
         :return: Return True if we want to consume this event so it is not passed on to the
                  rest of the UI.
         """
-        if (event.type == pygame.USEREVENT and event.user_type == UI_BUTTON_PRESSED and
-                event.ui_element in [self.open_button, self.selected_option_button]):
-
+        if event.type == UI_BUTTON_PRESSED and event.ui_element in self.active_buttons:
             self.should_transition = True
 
         return False
@@ -531,12 +550,13 @@ class UIClosedDropDownState:
         self.base_position_rect.y = self.drop_down_menu_ui.relative_rect.y + border_and_shadow
 
         # update all the ui elements that depend on the base position
-        self.selected_option_button.set_relative_position(self.base_position_rect.topleft)
+        self.selected_option_button.set_relative_position((border_and_shadow, border_and_shadow))
 
-        open_button_x = (self.base_position_rect.x +
-                         self.base_position_rect.width -
-                         self.open_button_width)
-        self.open_button.set_relative_position((open_button_x, self.base_position_rect.y))
+        if self.open_button is not None:
+            open_button_x = (border_and_shadow +
+                             self.base_position_rect.width -
+                             self.open_button_width)
+            self.open_button.set_relative_position((open_button_x, self.base_position_rect.y))
 
     def update_dimensions(self):
         """
@@ -557,11 +577,13 @@ class UIClosedDropDownState:
         self.selected_option_button.set_dimensions((self.base_position_rect.width -
                                                     self.open_button_width,
                                                     self.base_position_rect.height))
-        open_button_x = (self.base_position_rect.x +
-                         self.base_position_rect.width -
-                         self.open_button_width)
-        self.open_button.set_dimensions((self.open_button_width, self.base_position_rect.height))
-        self.open_button.set_relative_position((open_button_x, self.base_position_rect.y))
+        if self.open_button is not None:
+            open_button_x = (border_and_shadow +
+                             self.base_position_rect.width -
+                             self.open_button_width)
+            self.open_button.set_dimensions((self.open_button_width,
+                                             self.base_position_rect.height))
+            self.open_button.set_relative_position((open_button_x, border_and_shadow))
 
     def on_fresh_drawable_shape_ready(self):
         """
@@ -595,7 +617,7 @@ class UIClosedDropDownState:
             self.selected_option_button.hide()
 
 
-class UIDropDownMenu(UIElement):
+class UIDropDownMenu(UIContainer):
     """
     A drop down menu lets us choose one text option from a list. That list of options can be
     expanded and hidden at the press of a button. While the element is called a drop down,
@@ -633,8 +655,8 @@ class UIDropDownMenu(UIElement):
                  visible: int = 1
                  ):
 
-        super().__init__(relative_rect, manager, container,
-                         layer_thickness=3, starting_height=1,
+        super().__init__(relative_rect, manager, container=container,
+                         starting_height=0,
                          anchors=anchors,
                          visible=visible)
 
@@ -665,16 +687,14 @@ class UIDropDownMenu(UIElement):
         self.expand_direction = None
 
         self.menu_states = {}
-
         self.rebuild_from_changed_theme_data()
-
         self.menu_states = {'closed': UIClosedDropDownState(self,
                                                             self.selected_option,
                                                             self.background_rect,
                                                             self.open_button_width,
                                                             self.expand_direction,
                                                             self.ui_manager,
-                                                            self.ui_container,
+                                                            self,
                                                             self.element_ids,
                                                             self.object_ids,
                                                             self.visible),
@@ -685,12 +705,12 @@ class UIDropDownMenu(UIElement):
                                                                 self.open_button_width,
                                                                 self.expand_direction,
                                                                 self.ui_manager,
-                                                                self.ui_container,
+                                                                self,
                                                                 self.element_ids,
                                                                 self.object_ids
                                                                 )}
         self.current_state = self.menu_states['closed']
-        self.current_state.start()
+        self.current_state.start(should_rebuild=True)
 
     def kill(self):
         """
@@ -760,6 +780,11 @@ class UIDropDownMenu(UIElement):
                                                                'rounded_rectangle']):
             has_any_changed = True
 
+        if self._check_misc_theme_data_changed(attribute_name='open_button_width',
+                                               default_value=20,
+                                               casting_func=int):
+            has_any_changed = True
+
         if self._check_shape_theming_changed(defaults={'border_width': 1,
                                                        'shadow_width': 2,
                                                        'shape_corner_radius': 2}):
@@ -796,8 +821,8 @@ class UIDropDownMenu(UIElement):
                                                self.relative_rect.width - (2 * border_and_shadow),
                                                self.relative_rect.height - (2 * border_and_shadow))
 
-            for state in self.menu_states:
-                self.menu_states[state].expand_direction = self.expand_direction
+            for _, state in self.menu_states.items():
+                state.expand_direction = self.expand_direction
 
             self.rebuild()
 
