@@ -141,7 +141,12 @@ class UIScreenSpaceHealthBar(UIElement):
                               'font': self.font,
                               'text': health_display_string,
                               'normal_text': self.text_colour,
-                              'text_shadow': self.text_shadow_colour,
+                              'normal_text_shadow': self.text_shadow_colour,
+                              'text_shadow': (1,
+                                              0,
+                                              0,
+                                              self.text_shadow_colour,
+                                              False),
                               'text_horiz_alignment': self.text_horiz_alignment,
                               'text_vert_alignment': self.text_vert_alignment,
                               'text_horiz_alignment_padding': self.text_horiz_alignment_padding,
