@@ -1,26 +1,33 @@
 import warnings
 import math
 
-from typing import Union, Tuple, Dict
+from typing import Union, Tuple, Dict, Optional, Any
 
 import pygame
 
 from pygame_gui.core import ObjectID
-from pygame_gui._constants import UI_TEXT_BOX_LINK_CLICKED
-from pygame_gui._constants import TEXT_EFFECT_TYPING_APPEAR
-from pygame_gui._constants import TEXT_EFFECT_FADE_IN, TEXT_EFFECT_FADE_OUT
+from pygame_gui._constants import UI_TEXT_BOX_LINK_CLICKED, OldType, UITextEffectType
+from pygame_gui._constants import TEXT_EFFECT_TYPING_APPEAR, TEXT_EFFECT_TILT
+from pygame_gui._constants import TEXT_EFFECT_FADE_IN, TEXT_EFFECT_FADE_OUT, TEXT_EFFECT_BOUNCE
+from pygame_gui._constants import TEXT_EFFECT_EXPAND_CONTRACT
 
+from pygame_gui.core.utility import translate
 from pygame_gui.core.interfaces import IContainerLikeInterface, IUIManagerInterface
+from pygame_gui.core.interfaces import IUITextOwnerInterface
 from pygame_gui.core.ui_element import UIElement
 from pygame_gui.core.drawable_shapes import RectDrawableShape, RoundedRectangleShape
 from pygame_gui.core.utility import basic_blit
 
 from pygame_gui.elements.ui_vertical_scroll_bar import UIVerticalScrollBar
-from pygame_gui.elements.text import TextBlock, TextHTMLParser
-from pygame_gui.elements.text import TypingAppearEffect, FadeInEffect, FadeOutEffect
+
+from pygame_gui.core.text.html_parser import HTMLParser
+from pygame_gui.core.text.text_box_layout import TextBoxLayout
+from pygame_gui.core.text.text_line_chunk import TextLineChunkFTFont
+from pygame_gui.core.text.text_effects import TypingAppearEffect, FadeInEffect, FadeOutEffect
+from pygame_gui.core.text.text_effects import BounceEffect, TiltEffect, ExpandContractEffect
 
 
-class UITextBox(UIElement):
+class UITextBox(UIElement, IUITextOwnerInterface):
     """
     A Text Box element lets us display word-wrapped, formatted text. If the text to display is
     longer than the height of the box given then the element will automatically create a vertical
@@ -86,12 +93,14 @@ class UITextBox(UIElement):
                                element_id='text_box')
 
         self.html_text = html_text
+        self.appended_text = ""
         self.font_dict = self.ui_theme.get_font_dictionary()
 
         self.wrap_to_height = wrap_to_height
         self.link_hover_chunks = []  # container for any link chunks we have
 
         self.active_text_effect = None
+        self.active_text_chunk_effects = []
         self.scroll_bar = None
         self.scroll_bar_width = 20
 
@@ -109,17 +118,24 @@ class UITextBox(UIElement):
         self.link_style = None
 
         self.rounded_corner_offset = None
-        self.formatted_text_block = None  # TextLine()
-        self.text_wrap_rect = None
+        self.text_box_layout = None  # type: Optional[TextBoxLayout]
+        self.text_wrap_rect = None  # type: Optional[pygame.Rect]
         self.background_surf = None
 
         self.drawable_shape = None
         self.shape = 'rectangle'
         self.shape_corner_radius = None
 
+        self.text_horiz_alignment = 'default'
+        self.text_vert_alignment = 'default'
+        self.text_horiz_alignment_padding = 0
+        self.text_vert_alignment_padding = 0
+
         self.should_trigger_full_rebuild = True
         self.time_until_full_rebuild_after_changing_size = 0.2
         self.full_rebuild_countdown = self.time_until_full_rebuild_after_changing_size
+
+        self.parser = None
 
         self.rebuild_from_changed_theme_data()
 
@@ -147,38 +163,38 @@ class UITextBox(UIElement):
         self.rounded_corner_offset = int(self.shape_corner_radius -
                                          (math.sin(math.pi / 4) *
                                           self.shape_corner_radius))
-        self.text_wrap_rect = [(self.rect[0] +
-                                self.padding[0] +
-                                self.border_width +
-                                self.shadow_width +
-                                self.rounded_corner_offset),
-                               (self.rect[1] +
-                                self.padding[1] +
-                                self.border_width +
-                                self.shadow_width +
-                                self.rounded_corner_offset),
-                               max(1, (self.rect[2] -
-                                       (self.padding[0] * 2) -
-                                       (self.border_width * 2) -
-                                       (self.shadow_width * 2) -
-                                       (2 * self.rounded_corner_offset))),
-                               max(1, (self.rect[3] -
-                                       (self.padding[1] * 2) -
-                                       (self.border_width * 2) -
-                                       (self.shadow_width * 2) -
-                                       (2 * self.rounded_corner_offset)))]
+        self.text_wrap_rect = pygame.Rect((self.rect[0] +
+                                           self.padding[0] +
+                                           self.border_width +
+                                           self.shadow_width +
+                                           self.rounded_corner_offset),
+                                          (self.rect[1] +
+                                           self.padding[1] +
+                                           self.border_width +
+                                           self.shadow_width +
+                                           self.rounded_corner_offset),
+                                          max(1, (self.rect[2] -
+                                                  (self.padding[0] * 2) -
+                                                  (self.border_width * 2) -
+                                                  (self.shadow_width * 2) -
+                                                  (2 * self.rounded_corner_offset))),
+                                          max(1, (self.rect[3] -
+                                                  (self.padding[1] * 2) -
+                                                  (self.border_width * 2) -
+                                                  (self.shadow_width * 2) -
+                                                  (2 * self.rounded_corner_offset))))
         if self.wrap_to_height or self.rect[3] == -1:
-            self.text_wrap_rect[3] = -1
+            self.text_wrap_rect.height = -1
         if self.rect[2] == -1:
-            self.text_wrap_rect[2] = -1
+            self.text_wrap_rect.width = -1
 
         drawable_area_size = (self.text_wrap_rect[2], self.text_wrap_rect[3])
 
         # This gives us the height of the text at the 'width' of the text_wrap_area
         self.parse_html_into_style_data()
-        if self.formatted_text_block is not None:
+        if self.text_box_layout is not None:
             if self.wrap_to_height or self.rect[3] == -1 or self.rect[2] == -1:
-                final_text_area_size = self.formatted_text_block.final_dimensions
+                final_text_area_size = self.text_box_layout.layout_rect.size
                 new_dimensions = ((final_text_area_size[0] + (self.padding[0] * 2) +
                                    (self.border_width * 2) + (self.shadow_width * 2) +
                                    (2 * self.rounded_corner_offset)),
@@ -199,7 +215,7 @@ class UITextBox(UIElement):
                                               (self.shadow_width * 2) -
                                               (2 * self.rounded_corner_offset))))
 
-            elif self.formatted_text_block.final_dimensions[1] > self.text_wrap_rect[3]:
+            elif self.text_box_layout.layout_rect.height > self.text_wrap_rect[3]:
                 # We need a scrollbar because our text is longer than the space we
                 # have to display it. This also means we need to parse the text again.
                 text_rect_width = (self.rect[2] -
@@ -207,19 +223,22 @@ class UITextBox(UIElement):
                                    (self.border_width * 2) -
                                    (self.shadow_width * 2) -
                                    self.rounded_corner_offset - self.scroll_bar_width)
-                self.text_wrap_rect = [(self.rect[0] + self.padding[0] + self.border_width +
-                                        self.shadow_width + self.rounded_corner_offset),
-                                       (self.rect[1] + self.padding[1] + self.border_width +
-                                        self.shadow_width + self.rounded_corner_offset),
-                                       max(1, text_rect_width),
-                                       max(1, (self.rect[3] -
-                                               (self.padding[1] * 2) -
-                                               (self.border_width * 2) -
-                                               (self.shadow_width * 2) -
-                                               (2 * self.rounded_corner_offset)))]
+                self.text_wrap_rect = pygame.Rect((self.rect[0] + self.padding[0] +
+                                                   self.border_width + self.shadow_width +
+                                                   self.rounded_corner_offset),
+                                                  (self.rect[1] + self.padding[1] +
+                                                   self.border_width +
+                                                   self.shadow_width +
+                                                   self.rounded_corner_offset),
+                                                  max(1, text_rect_width),
+                                                  max(1, (self.rect[3] -
+                                                          (self.padding[1] * 2) -
+                                                          (self.border_width * 2) -
+                                                          (self.shadow_width * 2) -
+                                                          (2 * self.rounded_corner_offset))))
                 self.parse_html_into_style_data()
                 percentage_visible = (self.text_wrap_rect[3] /
-                                      self.formatted_text_block.final_dimensions[1])
+                                      self.text_box_layout.layout_rect.height)
                 scroll_bar_position = (self.relative_rect.right - self.border_width -
                                        self.shadow_width - self.scroll_bar_width,
                                        self.relative_rect.top + self.border_width +
@@ -258,7 +277,7 @@ class UITextBox(UIElement):
 
         if self.scroll_bar is not None:
             height_adjustment = int(self.scroll_bar.start_percentage *
-                                    self.formatted_text_block.final_dimensions[1])
+                                    self.text_box_layout.layout_rect.height)
         else:
             height_adjustment = 0
 
@@ -270,7 +289,8 @@ class UITextBox(UIElement):
         new_image = pygame.surface.Surface(self.rect.size, flags=pygame.SRCALPHA, depth=32)
         new_image.fill(pygame.Color(0, 0, 0, 0))
         basic_blit(new_image, self.background_surf, (0, 0))
-        basic_blit(new_image, self.formatted_text_block.block_sprite,
+
+        basic_blit(new_image, self.text_box_layout.finalised_surface,
                    (self.padding[0] + self.border_width +
                     self.shadow_width + self.rounded_corner_offset,
                     self.padding[1] + self.border_width +
@@ -278,11 +298,34 @@ class UITextBox(UIElement):
                    drawable_area)
 
         self.set_image(new_image)
-
-        self.formatted_text_block.add_chunks_to_hover_group(self.link_hover_chunks)
+        self.link_hover_chunks = []
+        self.text_box_layout.add_chunks_to_hover_group(self.link_hover_chunks)
 
         self.should_trigger_full_rebuild = False
         self.full_rebuild_countdown = self.time_until_full_rebuild_after_changing_size
+
+    def _align_all_text_rows(self):
+        """
+        Aligns the text drawing position correctly according to our theming options.
+
+        """
+        # Horizontal alignment
+        if self.text_horiz_alignment != 'default':
+            if self.text_horiz_alignment == 'center':
+                self.text_box_layout.horiz_center_all_rows()
+            elif self.text_horiz_alignment == 'left':
+                self.text_box_layout.align_left_all_rows(self.text_horiz_alignment_padding)
+            else:
+                self.text_box_layout.align_right_all_rows(self.text_horiz_alignment_padding)
+
+        # Vertical alignment
+        if self.text_vert_alignment != 'default':
+            if self.text_vert_alignment == 'center':
+                self.text_box_layout.vert_center_all_rows()
+            elif self.text_vert_alignment == 'top':
+                self.text_box_layout.vert_align_top_all_rows(self.text_vert_alignment_padding)
+            else:
+                self.text_box_layout.vert_align_bottom_all_rows(self.text_vert_alignment_padding)
 
     def update(self, time_delta: float):
         """
@@ -298,7 +341,7 @@ class UITextBox(UIElement):
             return
         if self.scroll_bar is not None and self.scroll_bar.check_has_moved_recently():
             height_adjustment = int(self.scroll_bar.start_percentage *
-                                    self.formatted_text_block.final_dimensions[1])
+                                    self.text_box_layout.layout_rect.height)
 
             drawable_area_size = (max(1, (self.rect[2] -
                                           (self.padding[0] * 2) -
@@ -319,7 +362,7 @@ class UITextBox(UIElement):
             new_image = pygame.surface.Surface(self.rect.size, flags=pygame.SRCALPHA, depth=32)
             new_image.fill(pygame.Color(0, 0, 0, 0))
             basic_blit(new_image, self.background_surf, (0, 0))
-            basic_blit(new_image, self.formatted_text_block.block_sprite,
+            basic_blit(new_image, self.text_box_layout.finalised_surface,
                        (self.padding[0] + self.border_width +
                         self.shadow_width +
                         self.rounded_corner_offset,
@@ -330,11 +373,11 @@ class UITextBox(UIElement):
             self.set_image(new_image)
 
         mouse_x, mouse_y = self.ui_manager.get_mouse_position()
-        should_redraw_from_chunks = False
+        should_redraw_from_layout = False
 
         if self.scroll_bar is not None:
             height_adjustment = (self.scroll_bar.start_percentage *
-                                 self.formatted_text_block.final_dimensions[1])
+                                 self.text_box_layout.layout_rect.height)
         else:
             height_adjustment = 0
         base_x = int(self.rect[0] + self.padding[0] + self.border_width +
@@ -345,28 +388,23 @@ class UITextBox(UIElement):
         for chunk in self.link_hover_chunks:
             hovered_currently = False
 
-            hover_rect = pygame.Rect((base_x + chunk.rect.x,
-                                      base_y + chunk.rect.y),
-                                     chunk.rect.size)
+            hover_rect = pygame.Rect((base_x + chunk.x,
+                                      base_y + chunk.y),
+                                     chunk.size)
             if hover_rect.collidepoint(mouse_x, mouse_y) and self.rect.collidepoint(mouse_x,
                                                                                     mouse_y):
                 hovered_currently = True
             if chunk.is_hovered and not hovered_currently:
                 chunk.on_unhovered()
-                should_redraw_from_chunks = True
+                should_redraw_from_layout = True
             elif hovered_currently and not chunk.is_hovered:
                 chunk.on_hovered()
-                should_redraw_from_chunks = True
+                should_redraw_from_layout = True
 
-        if should_redraw_from_chunks:
-            self.redraw_from_chunks()
+        if should_redraw_from_layout:
+            self.redraw_from_text_block()
 
-        if self.active_text_effect is not None:
-            self.active_text_effect.update(time_delta)
-            if self.active_text_effect.should_full_redraw():
-                self.full_redraw()
-            if self.active_text_effect.should_redraw_from_chunks():
-                self.redraw_from_chunks()
+        self.update_text_effect(time_delta)
 
         if self.should_trigger_full_rebuild and self.full_rebuild_countdown <= 0.0:
             self.rebuild()
@@ -471,21 +509,24 @@ class UITextBox(UIElement):
 
     def parse_html_into_style_data(self):
         """
-        Parses HTML styled string text into a format more useful for styling pygame.font
+        Parses HTML styled string text into a format more useful for styling pygame.freetype
         rendered text.
         """
-        parser = TextHTMLParser(self.ui_theme, self.combined_element_ids)
-        parser.push_style('body', {"bg_colour": self.background_colour})
-        parser.feed(self.html_text)
 
-        self.formatted_text_block = TextBlock(parser.text_data,
-                                              self.text_wrap_rect,
-                                              parser.indexed_styles,
-                                              self.font_dict,
-                                              self.link_style,
-                                              self.background_colour,
-                                              self.wrap_to_height
-                                              )
+        self.parser.feed(translate(self.html_text) + self.appended_text)
+
+        self.text_box_layout = TextBoxLayout(self.parser.layout_rect_queue,
+                                             pygame.Rect((0, 0), (self.text_wrap_rect[2],
+                                                                  self.text_wrap_rect[3])),
+                                             pygame.Rect((0, 0), (self.text_wrap_rect[2],
+                                                                  self.text_wrap_rect[3])),
+                                             line_spacing=1.25)
+        self.parser.empty_layout_queue()
+        if self.text_wrap_rect[3] == -1:
+            self.text_box_layout.view_rect.height = self.text_box_layout.layout_rect.height
+
+        self._align_all_text_rows()
+        self.text_box_layout.finalise_to_new()
 
     def redraw_from_text_block(self):
         """
@@ -497,7 +538,10 @@ class UITextBox(UIElement):
             return
         if self.scroll_bar is not None:
             height_adjustment = int(self.scroll_bar.start_percentage *
-                                    self.formatted_text_block.final_dimensions[1])
+                                    self.text_box_layout.layout_rect.height)
+            percentage_visible = (self.text_wrap_rect[3] /
+                                  self.text_box_layout.layout_rect.height)
+            self.scroll_bar.set_visible_percentage(percentage_visible)
         else:
             height_adjustment = 0
         drawable_area_size = (max(1, (self.rect[2] -
@@ -515,7 +559,7 @@ class UITextBox(UIElement):
         new_image = pygame.surface.Surface(self.rect.size, flags=pygame.SRCALPHA, depth=32)
         new_image.fill(pygame.Color(0, 0, 0, 0))
         basic_blit(new_image, self.background_surf, (0, 0))
-        basic_blit(new_image, self.formatted_text_block.block_sprite,
+        basic_blit(new_image, self.text_box_layout.finalised_surface,
                    (self.padding[0] + self.border_width +
                     self.shadow_width + self.rounded_corner_offset,
                     self.padding[1] + self.border_width +
@@ -532,7 +576,7 @@ class UITextBox(UIElement):
 
         This won't work very well if redrawing a chunk changed it's dimensions.
         """
-        self.formatted_text_block.redraw_from_chunks(self.active_text_effect)
+        self.text_box_layout.finalise_to_new()
         self.redraw_from_text_block()
 
     def full_redraw(self):
@@ -545,10 +589,13 @@ class UITextBox(UIElement):
         new text box.
 
         """
-        self.formatted_text_block.redraw(self.active_text_effect)
+        self.text_box_layout.reprocess_layout_queue(pygame.Rect((0, 0),
+                                                                (self.text_wrap_rect[2],
+                                                                 self.text_wrap_rect[3])))
+        self.text_box_layout.finalise_to_new()
         self.redraw_from_text_block()
         self.link_hover_chunks = []
-        self.formatted_text_block.add_chunks_to_hover_group(self.link_hover_chunks)
+        self.text_box_layout.add_chunks_to_hover_group(self.link_hover_chunks)
 
     def process_event(self, event: pygame.event.Event) -> bool:
         """
@@ -560,8 +607,7 @@ class UITextBox(UIElement):
 
         """
         consumed_event = False
-        should_redraw_from_chunks = False
-        should_full_redraw = False
+        should_redraw_from_layout = False
         if event.type == pygame.MOUSEBUTTONDOWN and event.button == 1:
             scaled_mouse_pos = self.ui_manager.calculate_scaled_mouse_position(event.pos)
             if self.hover_point(scaled_mouse_pos[0], scaled_mouse_pos[1]):
@@ -569,7 +615,7 @@ class UITextBox(UIElement):
 
                 if self.is_enabled:
                     if self.scroll_bar is not None:
-                        text_block_full_height = self.formatted_text_block.final_dimensions[1]
+                        text_block_full_height = self.text_box_layout.layout_rect.height
                         height_adjustment = (self.scroll_bar.start_percentage *
                                              text_block_full_height)
                     else:
@@ -580,22 +626,19 @@ class UITextBox(UIElement):
                                  self.shadow_width + self.rounded_corner_offset - height_adjustment)
                     for chunk in self.link_hover_chunks:
 
-                        hover_rect = pygame.Rect((base_x + chunk.rect.x,
-                                                  base_y + chunk.rect.y),
-                                                 chunk.rect.size)
+                        hover_rect = pygame.Rect((base_x + chunk.x,
+                                                  base_y + chunk.y),
+                                                 chunk.size)
                         if hover_rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1]):
                             consumed_event = True
-                            if not chunk.is_selected:
-                                chunk.on_selected()
-                                if chunk.metrics_changed_after_redraw:
-                                    should_full_redraw = True
-                                else:
-                                    should_redraw_from_chunks = True
+                            if not chunk.is_active:
+                                chunk.set_active()
+                                should_redraw_from_layout = True
 
         if self.is_enabled and event.type == pygame.MOUSEBUTTONUP and event.button == 1:
             if self.scroll_bar is not None:
                 height_adjustment = (self.scroll_bar.start_percentage *
-                                     self.formatted_text_block.final_dimensions[1])
+                                     self.text_box_layout.layout_rect.height)
             else:
                 height_adjustment = 0
             base_x = int(self.rect[0] + self.padding[0] + self.border_width +
@@ -605,66 +648,35 @@ class UITextBox(UIElement):
             scaled_mouse_pos = self.ui_manager.calculate_scaled_mouse_position(event.pos)
             for chunk in self.link_hover_chunks:
 
-                hover_rect = pygame.Rect((base_x + chunk.rect.x,
-                                          base_y + chunk.rect.y),
-                                         chunk.rect.size)
+                hover_rect = pygame.Rect((base_x + chunk.x,
+                                          base_y + chunk.y),
+                                         chunk.size)
                 if (hover_rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1]) and
                         self.rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1])):
                     consumed_event = True
-                    if chunk.is_selected:
-                        event_data = {'user_type': UI_TEXT_BOX_LINK_CLICKED,
-                                      'link_target': chunk.link_href,
+                    if chunk.is_active:
+
+                        # old event - to be removed in 0.8.0
+                        event_data = {'user_type': OldType(UI_TEXT_BOX_LINK_CLICKED),
+                                      'link_target': chunk.href,
                                       'ui_element': self,
                                       'ui_object_id': self.most_specific_combined_id}
                         pygame.event.post(pygame.event.Event(pygame.USEREVENT, event_data))
 
-                if chunk.is_selected:
-                    chunk.on_unselected()
-                    if chunk.metrics_changed_after_redraw:
-                        should_full_redraw = True
-                    else:
-                        should_redraw_from_chunks = True
+                        # new event
+                        event_data = {'link_target': chunk.href,
+                                      'ui_element': self,
+                                      'ui_object_id': self.most_specific_combined_id}
+                        pygame.event.post(pygame.event.Event(UI_TEXT_BOX_LINK_CLICKED, event_data))
 
-        if should_redraw_from_chunks:
-            self.redraw_from_chunks()
+                if chunk.is_active:
+                    chunk.set_inactive()
+                    should_redraw_from_layout = True
 
-        if should_full_redraw:
-            self.full_redraw()
+        if should_redraw_from_layout:
+            self.redraw_from_text_block()
 
         return consumed_event
-
-    def set_active_effect(self, effect_name: Union[str, None]):
-        """
-        Set an animation effect to run on the text box. The effect will start running immediately
-        after this call.
-
-        These effects are currently supported:
-
-        - 'typing_appear' - Will look as if the text is being typed in.
-        - 'fade_in' - The text will fade in from the background colour (Only supported on Pygame 2)
-        - 'fade_out' - The text will fade out to the background colour (only supported on Pygame 2)
-
-        :param effect_name: The name fo the t to set. If set to None instead it will cancel any
-                            active effect.
-
-        """
-        if effect_name is None:
-            self.active_text_effect = None
-        elif isinstance(effect_name, str):
-            if effect_name == TEXT_EFFECT_TYPING_APPEAR:
-                effect = TypingAppearEffect(self.formatted_text_block.characters)
-                self.active_text_effect = effect
-                self.full_redraw()
-            elif effect_name == TEXT_EFFECT_FADE_IN:
-                effect = FadeInEffect(self.formatted_text_block.characters)
-                self.active_text_effect = effect
-                self.redraw_from_chunks()
-            elif effect_name == TEXT_EFFECT_FADE_OUT:
-                effect = FadeOutEffect(self.formatted_text_block.characters)
-                self.active_text_effect = effect
-                self.redraw_from_chunks()
-            else:
-                warnings.warn('Unsupported effect name: ' + effect_name + ' for text box')
 
     def rebuild_from_changed_theme_data(self):
         """
@@ -708,11 +720,50 @@ class UITextBox(UIElement):
             self.border_colour = border_colour
             has_any_changed = True
 
+        if self._check_text_alignment_theming():
+            has_any_changed = True
+
         if self._check_link_style_changed():
             has_any_changed = True
 
         if has_any_changed:
-            self.rebuild()
+            self._reparse_and_rebuild()
+
+    def _reparse_and_rebuild(self):
+        self.parser = HTMLParser(self.ui_theme, self.combined_element_ids,
+                                 self.link_style, line_spacing=1.25)
+        self.rebuild()
+
+    def _check_text_alignment_theming(self) -> bool:
+        """
+        Checks for any changes in the theming data related to text alignment.
+
+        :return: True if changes found.
+
+        """
+        has_any_changed = False
+
+        if self._check_misc_theme_data_changed(attribute_name='text_horiz_alignment',
+                                               default_value='left',
+                                               casting_func=str):
+            has_any_changed = True
+
+        if self._check_misc_theme_data_changed(attribute_name='text_horiz_alignment_padding',
+                                               default_value=0,
+                                               casting_func=int):
+            has_any_changed = True
+
+        if self._check_misc_theme_data_changed(attribute_name='text_vert_alignment',
+                                               default_value='top',
+                                               casting_func=str):
+            has_any_changed = True
+
+        if self._check_misc_theme_data_changed(attribute_name='text_vert_alignment_padding',
+                                               default_value=0,
+                                               casting_func=int):
+            has_any_changed = True
+
+        return has_any_changed
 
     def _check_link_style_changed(self) -> bool:
         """
@@ -790,3 +841,200 @@ class UITextBox(UIElement):
 
         if self.scroll_bar is not None:
             self.scroll_bar.hide()
+
+    def append_html_text(self, new_html_str: str):
+        """
+        Adds a string, that is parsed for any HTML tags that pygame_gui supports, onto the bottom
+        of the text box's current contents.
+
+        This is useful for making things like logs.
+
+        :param new_html_str: The, potentially HTML tag, containing string of text to append.
+        """
+        self.appended_text += new_html_str
+        self.parser.feed(new_html_str)
+        self.text_box_layout.append_layout_rects(self.parser.layout_rect_queue)
+        self.parser.empty_layout_queue()
+
+        if (self.scroll_bar is None and
+                (self.text_box_layout.layout_rect.height > self.text_wrap_rect[3])):
+            self.rebuild()
+        else:
+            if self.scroll_bar is not None:
+                # set the scroll bar to the bottom
+                percentage_visible = (self.text_wrap_rect[3] /
+                                      self.text_box_layout.layout_rect.height)
+                self.scroll_bar.start_percentage = 1.0 - percentage_visible
+                self.scroll_bar.scroll_position = (self.scroll_bar.start_percentage *
+                                                   self.scroll_bar.scrollable_height)
+            self.redraw_from_text_block()
+
+    def on_locale_changed(self):
+        self._reparse_and_rebuild()
+
+    # -------------------------------------------------
+    # The Text owner interface
+    # -------------------------------------------------
+    def set_text_alpha(self, alpha: int, sub_chunk: Optional[TextLineChunkFTFont]):
+        if sub_chunk is None:
+            self.text_box_layout.set_alpha(alpha)
+        else:
+            sub_chunk.set_alpha(alpha)
+
+    def set_text_offset_pos(self, offset: Tuple[int, int],
+                            sub_chunk: Optional[TextLineChunkFTFont]):
+        if sub_chunk is None:
+            pass
+        else:
+            sub_chunk.set_offset_pos(offset)
+
+    def set_text_rotation(self, rotation: int,
+                          sub_chunk: Optional[TextLineChunkFTFont]):
+        if sub_chunk is None:
+            pass
+        else:
+            sub_chunk.set_rotation(rotation)
+
+    def set_text_scale(self, scale: float, sub_chunk: Optional[TextLineChunkFTFont]):
+        if sub_chunk is None:
+            pass
+        else:
+            sub_chunk.set_scale(scale)
+
+    def clear_text_surface(self, sub_chunk: Optional[TextLineChunkFTFont]):
+        if sub_chunk is None:
+            self.text_box_layout.clear_final_surface()
+        else:
+            sub_chunk.clear()
+
+    def get_text_letter_count(self, sub_chunk: Optional[TextLineChunkFTFont]) -> int:
+        if sub_chunk is None:
+            return self.text_box_layout.letter_count
+        else:
+            return sub_chunk.letter_count
+
+    def update_text_end_position(self, end_pos: int, sub_chunk: Optional[TextLineChunkFTFont]):
+        if sub_chunk is None:
+            self.text_box_layout.update_text_with_new_text_end_pos(end_pos)
+        else:
+            sub_chunk.letter_end = end_pos
+            sub_chunk.redraw()
+
+    def set_active_effect(self, effect_type: Optional[UITextEffectType],
+                          params: Optional[Dict[str, Any]] = None,
+                          effect_tag: Optional[str] = None):
+        if effect_tag is not None:
+            redrew_all_chunks = False
+            if self.active_text_effect is not None:
+                # don't allow mixing of full text box effects and chunk effects
+                self.clear_all_active_effects()
+                self.active_text_effect = None
+                redrew_all_chunks = True
+            # we have a tag so only want to apply our effect to tagged chunks
+            # first see if we have any tagged chunks in the layout
+            for row in self.text_box_layout.layout_rows:
+                for chunk in row.items:
+                    if isinstance(chunk, TextLineChunkFTFont):
+                        if chunk.effect_id == effect_tag:
+                            # need to clear off any old effects on this chunk too if we didn't
+                            # already redraw everything
+                            if not redrew_all_chunks:
+                                self.clear_all_active_effects(chunk)
+                            effect = None
+                            if effect_type == TEXT_EFFECT_TYPING_APPEAR:
+                                effect = TypingAppearEffect(self, params, chunk)
+                            elif effect_type == TEXT_EFFECT_FADE_IN:
+                                effect = FadeInEffect(self, params, chunk)
+                            elif effect_type == TEXT_EFFECT_FADE_OUT:
+                                effect = FadeOutEffect(self, params, chunk)
+                            elif effect_type == TEXT_EFFECT_BOUNCE:
+                                effect = BounceEffect(self, params, chunk)
+                            elif effect_type == TEXT_EFFECT_TILT:
+                                effect = TiltEffect(self, params, chunk)
+                            elif effect_type == TEXT_EFFECT_EXPAND_CONTRACT:
+                                effect = ExpandContractEffect(self, params, chunk)
+                            else:
+                                warnings.warn('Unsupported effect name: ' + str(
+                                    effect_type) + ' for text chunk')
+                            chunk.grab_pre_effect_surface()
+                            self.active_text_chunk_effects.append({'chunk': chunk,
+                                                                   'effect': effect})
+        else:
+            if self.active_text_effect is not None or len(self.active_text_chunk_effects) != 0:
+                self.clear_all_active_effects()
+            if effect_type is None:
+                self.active_text_effect = None
+            elif isinstance(effect_type, UITextEffectType):
+                if effect_type == TEXT_EFFECT_TYPING_APPEAR:
+                    self.active_text_effect = TypingAppearEffect(self, params)
+                elif effect_type == TEXT_EFFECT_FADE_IN:
+                    self.active_text_effect = FadeInEffect(self, params)
+                elif effect_type == TEXT_EFFECT_FADE_OUT:
+                    self.active_text_effect = FadeOutEffect(self, params)
+                else:
+                    warnings.warn('Unsupported effect name: '
+                                  + str(effect_type) + ' for whole text box')
+            else:
+                warnings.warn('Unsupported effect name: '
+                              + str(effect_type) + ' for whole text box')
+
+    def stop_finished_effect(self, sub_chunk: Optional[TextLineChunkFTFont] = None):
+        if sub_chunk is None:
+            self.active_text_effect = None
+        else:
+            self.active_text_chunk_effects = [effect_chunk for effect_chunk in
+                                              self.active_text_chunk_effects
+                                              if effect_chunk['chunk'] != sub_chunk]
+
+    def clear_all_active_effects(self, sub_chunk: Optional[TextLineChunkFTFont] = None):
+        if sub_chunk is None:
+            self.active_text_effect = None
+            self.text_box_layout.clear_effects()
+            for effect_chunk in self.active_text_chunk_effects:
+                effect_chunk['chunk'].clear_effects()
+            self.active_text_chunk_effects = []
+            self.text_box_layout.finalise_to_new()
+        else:
+            self.active_text_chunk_effects = [effect_chunk for effect_chunk in
+                                              self.active_text_chunk_effects
+                                              if effect_chunk['chunk'] != sub_chunk]
+            sub_chunk.clear_effects()
+
+            effect_chunks = []
+            for affected_chunk in self.active_text_chunk_effects:
+                effect_chunks.append(affected_chunk['chunk'])
+            self.text_box_layout.redraw_other_chunks(effect_chunks)
+
+            sub_chunk.redraw()
+
+    def update_text_effect(self, time_delta: float):
+        if self.active_text_effect is not None:
+            self.active_text_effect.update(time_delta)
+            # update can set effect to None
+            if (self.active_text_effect is not None and
+                    self.active_text_effect.has_text_changed()):
+                self.active_text_effect.apply_effect()
+                self.redraw_from_text_block()
+
+        if len(self.active_text_chunk_effects) > 0:
+            any_text_changed = False
+            for affected_chunk in self.active_text_chunk_effects:
+                affected_chunk['effect'].update(time_delta)
+                if (affected_chunk['effect'] is not None and
+                        affected_chunk['effect'].has_text_changed()):
+                    any_text_changed = True
+            if any_text_changed:
+                effect_chunks = []
+                for affected_chunk in self.active_text_chunk_effects:
+                    chunk = affected_chunk['chunk']
+                    chunk.clear(chunk.transform_effect_rect)
+                    effect_chunks.append(chunk)
+
+                self.text_box_layout.redraw_other_chunks(effect_chunks)
+
+                for affected_chunk in self.active_text_chunk_effects:
+                    affected_chunk['effect'].apply_effect()
+                self.redraw_from_text_block()
+
+    def get_object_id(self) -> str:
+        return self.most_specific_combined_id
