@@ -1,10 +1,12 @@
-from pygame_gui.windows.ui_message_window import UIMessageWindow
+from pygame_gui.windows.ui_message_window import UIMessageWindow  # noqa pylint: disable=cyclic-import
 from pygame_gui.windows.ui_confirmation_dialog import UIConfirmationDialog
 from pygame_gui.windows.ui_file_dialog import UIFileDialog
 from pygame_gui.windows.ui_colour_picker_dialog import UIColourPickerDialog
+from pygame_gui.windows.ui_console_window import UIConsoleWindow
 
 
 __all__ = ['UIMessageWindow',
            'UIConfirmationDialog',
            'UIFileDialog',
-           'UIColourPickerDialog']
+           'UIColourPickerDialog',
+           'UIConsoleWindow']
