@@ -1,7 +1,5 @@
+import pytest
 import pygame
-
-from tests.shared_fixtures import _init_pygame, default_ui_manager, default_display_surface, \
-    _display_surface_return_none
 
 from pygame_gui.core.drawable_shapes.rect_drawable_shape import RectDrawableShape
 from pygame_gui.ui_manager import UIManager
@@ -12,6 +10,8 @@ class TestRectDrawableShape:
         RectDrawableShape(containing_rect=pygame.Rect(0, 0, 100, 100),
                           theming_parameters={'text': 'test',
                                               'font': default_ui_manager.get_theme().get_font([]),
+                                              'normal_text': pygame.Color('#FFFFFF'),
+                                              'normal_text_shadow': pygame.Color('#000000'),
                                               'shadow_width': 0,
                                               'border_width': 0,
                                               'normal_border': pygame.Color('#FFFFFF'),
@@ -24,6 +24,8 @@ class TestRectDrawableShape:
         shape = RectDrawableShape(containing_rect=pygame.Rect(0, 0, 100, 100),
                                   theming_parameters={'text': 'test',
                                                       'font': default_ui_manager.get_theme().get_font([]),
+                                                      'normal_text': pygame.Color('#FFFFFF'),
+                                                      'normal_text_shadow': pygame.Color('#000000'),
                                                       'shadow_width': -10,
                                                       'border_width': -10,
                                                       'normal_border': pygame.Color('#FFFFFF'),
@@ -37,6 +39,8 @@ class TestRectDrawableShape:
         shape = RectDrawableShape(containing_rect=pygame.Rect(0, 0, 25, 25),
                                   theming_parameters={'text': 'test',
                                                       'font': default_ui_manager.get_theme().get_font([]),
+                                                      'normal_text': pygame.Color('#FFFFFF'),
+                                                      'normal_text_shadow': pygame.Color('#000000'),
                                                       'shadow_width': 20,
                                                       'border_width': 20,
                                                       'normal_border': pygame.Color('#FFFFFF'),
@@ -50,6 +54,8 @@ class TestRectDrawableShape:
         shape = RectDrawableShape(containing_rect=pygame.Rect(0, 0, 2, 2),
                                   theming_parameters={'text': 'test',
                                                       'font': default_ui_manager.get_theme().get_font([]),
+                                                      'normal_text': pygame.Color('#FFFFFF'),
+                                                      'normal_text_shadow': pygame.Color('#000000'),
                                                       'shadow_width': 1,
                                                       'border_width': 0,
                                                       'normal_border': pygame.Color('#FFFFFF'),
@@ -63,6 +69,8 @@ class TestRectDrawableShape:
         shape = RectDrawableShape(containing_rect=pygame.Rect(0, 0, 100, 100),
                                   theming_parameters={'text': 'test',
                                                       'font': default_ui_manager.get_theme().get_font([]),
+                                                      'normal_text': pygame.Color('#FFFFFF'),
+                                                      'normal_text_shadow': pygame.Color('#000000'),
                                                       'shadow_width': 0,
                                                       'border_width': 0,
                                                       'normal_border': pygame.Color('#FFFFFF'),
@@ -77,6 +85,8 @@ class TestRectDrawableShape:
         shape = RectDrawableShape(containing_rect=pygame.Rect(0, 0, 100, 100),
                                   theming_parameters={'text': 'test',
                                                       'font': default_ui_manager.get_theme().get_font([]),
+                                                      'normal_text': pygame.Color('#FFFFFF'),
+                                                      'normal_text_shadow': pygame.Color('#000000'),
                                                       'shadow_width': 0,
                                                       'border_width': 0,
                                                       'normal_border': pygame.Color('#FFFFFF'),
@@ -92,6 +102,8 @@ class TestRectDrawableShape:
         shape = RectDrawableShape(containing_rect=pygame.Rect(0, 0, 100, 100),
                                   theming_parameters={'text': 'test',
                                                       'font': default_ui_manager.get_theme().get_font([]),
+                                                      'normal_text': pygame.Color('#FFFFFF'),
+                                                      'normal_text_shadow': pygame.Color('#000000'),
                                                       'shadow_width': 0,
                                                       'border_width': 0,
                                                       'normal_border': pygame.Color('#FFFFFF'),
@@ -104,3 +116,5 @@ class TestRectDrawableShape:
         assert shape.containing_rect.width == 50
 
 
+if __name__ == '__main__':
+    pytest.console_main()
